Combine page create and edit forms into one component
