--- client/src/components/PageDisplay/PageEdit.vue
@@ -1,107 +1,0 @@
-<template>
-    <LoadingSpan v-if="loadingPage" />
-    <div v-else>
-        <BAlert v-if="errorMessage" variant="danger" show>{{ errorMessage }}</BAlert>
-        <FormCard title="Create a new Page" icon="fa-file-contract">
-            <template v-slot:body>
-                <FormElementLabel title="Title" :required="true" :condition="!!title">
-                    <FormInput id="page-title" v-model="title" />
-                </FormElementLabel>
-                <FormElementLabel
-                    title="Identifier"
-                    help="A unique identifier that will be used for public links to this page. This field can only contain lowercase letters, numbers, and dashes (-)."
-                    :required="true"
-                    :condition="!!slug">
-                    <FormInput id="page-slug" v-model="slug" />
-                </FormElementLabel>
-                <FormElementLabel
-                    title="Annotation"
-                    help="A description of the page. The annotation is shown alongside published pages.">
-                    <FormInput id="page-annotation" v-model="annotation" />
-                </FormElementLabel>
-            </template>
-        </FormCard>
-        <BButton id="page-submit" class="my-2" variant="primary" @click="onUpdate">
-            <FontAwesomeIcon :icon="faSave" class="mr-1" />
-            <span v-localize>Update</span>
-        </BButton>
-    </div>
-</template>
-
-<script setup lang="ts">
-import { faSave } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
-import { BAlert, BButton } from "bootstrap-vue";
-import { ref, watch } from "vue";
-import { useRouter } from "vue-router/composables";
-
-import { GalaxyApi } from "@/api";
-import pageTemplate from "@/components/PageDisplay/pageTemplate.yml";
-
-import FormInput from "@/components/Form/Elements/FormInput.vue";
-import FormCard from "@/components/Form/FormCard.vue";
-import FormElementLabel from "@/components/Form/FormElementLabel.vue";
-import LoadingSpan from "@/components/LoadingSpan.vue";
-
-const router = useRouter();
-
-const props = defineProps<{
-    id: string;
-}>();
-
-const annotation = ref("");
-const content = ref(pageTemplate.content);
-const errorMessage = ref("");
-const loadingPage = ref(false);
-const slug = ref("");
-const title = ref("");
-
-async function fetchPage() {
-    if (props.id) {
-        loadingPage.value = true;
-        const { data, error } = await GalaxyApi().GET("/api/pages/{id}", {
-            params: { path: { id: props.id } },
-        });
-        if (error) {
-            errorMessage.value = error.err_msg;
-        } else {
-            errorMessage.value = "";
-            annotation.value = data.annotation || "";
-            content.value = data.content;
-            slug.value = data.slug;
-            title.value = data.title;
-        }
-        loadingPage.value = false;
-    }
-}
-
-async function onUpdate() {
-    if (slug.value && title.value) {
-        const { error } = await GalaxyApi().PUT("/api/pages/{id}", {
-            params: {
-                path: {
-                    id: props.id,
-                }
-            },
-            body: {
-                annotation: annotation.value,
-                slug: slug.value,
-                title: title.value,
-            },
-        });
-        if (error) {
-            errorMessage.value = error.err_msg;
-        } else {
-            router.push(`/pages/list`);
-        }
-    } else {
-        errorMessage.value = "Please complete all required inputs.";
-    }
-}
-
-watch(
-    () => props.id,
-    () => fetchPage(),
-    { immediate: true }
-);
-</script>
